+++ rust/src/plane/mod.rs
@@ -183,8 +183,6 @@ mod reimplemented;
 pub(crate) use helpers::*;
 pub use reimplemented::*;
 
-use crate::NcDim;
-
 // NcPlane
 /// Fundamental drawing surface.
 ///
@@ -272,72 +270,3 @@ pub type NcFdPlane = crate::bindings::ffi::ncfdplane;
 ///
 /// `type in C: ncplane_options (struct)`
 pub type NcFdPlaneOptions = crate::bindings::ffi::ncfdplane_options;
-
-/// Blitter Mode (`NCBLIT_*`)
-///
-/// We never blit full blocks, but instead spaces (more efficient) with the
-/// background set to the desired foreground.
-///
-/// ## Modes
-///
-/// - [`NCBLIT_1x1`]
-/// - [`NCBLIT_2x1`]
-/// - [`NCBLIT_2x2`]
-/// - [`NCBLIT_3x2`]
-/// - [`NCBLIT_4x1`]
-/// - [`NCBLIT_8x1`]
-/// - [`NCBLIT_BRAILLE`]
-/// - [`NCBLIT_DEFAULT`]
-/// - [`NCBLIT_PIXEL`]
-///
-pub type NcBlitter = crate::bindings::ffi::ncblitter_e;
-
-/// [`NcBlitter`] mode using: space, compatible with ASCII
-pub const NCBLIT_1x1: NcBlitter = crate::bindings::ffi::ncblitter_e_NCBLIT_1x1;
-
-/// [`NcBlitter`] mode using: halves + 1x1 (space)
-/// ▄▀
-pub const NCBLIT_2x1: NcBlitter = crate::bindings::ffi::ncblitter_e_NCBLIT_2x1;
-
-/// [`NcBlitter`] mode using: quadrants + 2x1
-/// ▗▐ ▖▀▟▌▙
-pub const NCBLIT_2x2: NcBlitter = crate::bindings::ffi::ncblitter_e_NCBLIT_2x2;
-
-/// [`NcBlitter`] mode using: sextants
-/// 🬀🬁🬂🬃🬄🬅🬆🬇🬈🬉🬊🬋🬌🬍🬎🬏🬐🬑🬒🬓🬔🬕🬖🬗🬘🬙🬚🬛🬜🬝🬞🬟🬠🬡🬢🬣🬤🬥🬦🬧🬨🬩🬪🬫🬬🬭🬮🬯🬰🬱🬲🬳🬴🬵🬶🬷🬸🬹🬺🬻
-pub const NCBLIT_3x2: NcBlitter = crate::bindings::ffi::ncblitter_e_NCBLIT_3x2;
-
-/// [`NcBlitter`] mode using: four vertical levels
-/// █▆▄▂
-pub const NCBLIT_4x1: NcBlitter = crate::bindings::ffi::ncblitter_e_NCBLIT_4x1;
-
-/// [`NcBlitter`] mode using: eight vertical levels
-/// █▇▆▅▄▃▂▁
-pub const NCBLIT_8x1: NcBlitter = crate::bindings::ffi::ncblitter_e_NCBLIT_8x1;
-
-/// [`NcBlitter`] mode using: 4 rows, 2 cols (braille)
-/// ⡀⡄⡆⡇⢀⣀⣄⣆⣇⢠⣠⣤⣦⣧⢰⣰⣴⣶⣷⢸⣸⣼⣾⣿
-pub const NCBLIT_BRAILLE: NcBlitter = crate::bindings::ffi::ncblitter_e_NCBLIT_BRAILLE;
-
-/// [`NcBlitter`] mode where the blitter is automatically chosen
-pub const NCBLIT_DEFAULT: NcBlitter = crate::bindings::ffi::ncblitter_e_NCBLIT_DEFAULT;
-
-/// Sixel/Pixel mode
-///
-/// See [Sixel in Wikipedia](https://en.wikipedia.org/wiki/Sixel).
-pub const NCBLIT_PIXEL: NcBlitter = crate::bindings::ffi::ncblitter_e_NCBLIT_PIXEL;
-
-/// Contains the pixel geometry information as returned by the
-/// NcPlane.[pixelgeom()][NcPlane#method.pixelgeom] method.
-///
-/// If bitmaps are not supported, the fields `max_bitmap_*` will be 0.
-#[derive(Clone, Debug)]
-pub struct NcPixelGeometry {
-    /// Geometry of the display region
-    pub display_y: NcDim,
-    pub display_x: NcDim,
-    pub cell_y: NcDim,
-    pub cell_x: NcDim,
-    pub max_bitmap_y: NcDim,
-    pub max_bitmap_x: NcDim,
-}
